+++ controller/upcomingEvent.controller.ts
@@ -1,9 +1,19 @@
 import { Request, Response } from "express";
 import upcomingEventModel from "../model/upcomingEvent.model";
+import {
+  isDatabaseConnected,
+  sendDatabaseUnavailable,
+  sendEmptyListWhenDatabaseUnavailable,
+} from "../utils/databaseStatus";
 import { getErrorMessage } from "../utils/imageUpload";
 
 export const getUpcomingEvents = async (_: Request, res: Response) => {
   try {
+    if (!isDatabaseConnected()) {
+      sendEmptyListWhenDatabaseUnavailable(res);
+      return;
+    }
+
     const events = await upcomingEventModel.find({ isActive: true }).sort({ createdAt: -1 });
     res.status(200).json(events);
   } catch (error) {
@@ -14,6 +24,11 @@ export const getUpcomingEvents = async (_: Request, res: Response) => {
 
 export const createUpcomingEvent = async (req: Request, res: Response) => {
   try {
+    if (!isDatabaseConnected()) {
+      sendDatabaseUnavailable(res);
+      return;
+    }
+
     const { title } = req.body;
 
     if (!title) {
@@ -36,6 +51,11 @@ export const createUpcomingEvent = async (req: Request, res: Response) => {
 
 export const deleteUpcomingEvent = async (req: Request, res: Response) => {
   try {
+    if (!isDatabaseConnected()) {
+      sendDatabaseUnavailable(res);
+      return;
+    }
+
     const deletedEvent = await upcomingEventModel.findByIdAndDelete(req.params.id);
 
     if (!deletedEvent) {
